Improve awareness of replicas
Be aware of replicated/global mode and calculate the correct number of
desired replicas for any service (either the specified number in
replicated mode or the number of nodes in global mode). Display icon for
global/replicated mode.

General improvements to UI - improve (?) node heartbeat visualisation,
node stats layout and labelling, ordering of items in stack view for
relevance and consistency, account for "other" services (i.e. all those
not in a stack), modify Darkly default green colour as it was a bit
bright for my taste.