Create standalone service for harvesting swarm state
Previous design would harvest swarm state for every client. i.e. if 10
clients are active then the swarm state is harvested 10 times every 5
seconds or so. This can get very expensive on a busy swarm and results
in very slow web requests even for loading the main app.

Since we can't scale up our web server processes easily (due to a
restriction of flask_socketio limiting to only one worker process) it
makes more sense to move the harvesting code into a separate service
that does the expensive/slow processing at a given interval. This data
is sent to the web app which re-broadcasts it to all websocket clients
which is much more efficient.