Handle service discovery race condition
Monitor: Deal with case where a service is removed while we are
inspecting it.

Agent: Deal with case where a container is removed while we are
inspecting it.

Avoid UI transition loop from MouseLeave event - do not trigger timer if
overview is not visible.

Fix height of dashboard to remove scrollbar on main viewport.