Include both messages and state in container overview detail
Combine `message` and `state` field as they are very closely related and
sometimes contain the same data.

Fix issues with React keys causing incorrect element display.

In overview detail view, display rows as chunks of 4 rows side by side.
In the case of the Containers detail view, the data is dynamic so we
cannot restrict to 4 rows.