Convert timestamps on chart data, smooth values
Use Redux reducer to convert timestamps on provided chart data.

Fix "no data" recognition - display a useful message when there is not
enough data available to render a chart.

When reducing number of datapoints for a chart (limited to 500) use the
mean of each datapoint across the reduced range to provide a
representative dataset.