Allow node warnings to repeat but slow down animation
Provide an alternative `pulse-<STATE>-loop` animation for each state
(error, warning, success) for node icon which only animates for 20% of
the animation, leaving 80% blank. This means we can use a 5s loop to get
the same 1s animation and 4s of nothing. This is much less intrusive in
the UI.