Simplify container stats collection
Running N threads (N=number of containers running on a given node) to
harvest statistics of each container seems to cause some problems as the
agent often gets stuck, possibly in a deadlock or waiting forever for a
thread to join. Instead of trying to make the threaded solution more
stable, simplify collection instead by only collecting stats for a small
number of containers (default: 1) on each iteration. Container stats
will be updated less frequently but everything will run in a single
thread and the agent will update much more frequently which should
reduce false-positive notifications of a node being unreachable.

Reduce visual noise by using a green status indicator to identify a node
as healthy/reachable rather than pulsing yellow on each ping - users
only need to see when a node is having issues; the frequency of the
update is not particularly useful.