Handle chart build errors
Increase coverage of chart build error detection and ensure we always
report a failure to the front end rather than leaving the chart hanging
indefinitely.

Since we return the chart build request immediately and then build the
chart in the background, subsequently sending a chart response to update
the front end, we cannot simply wait for the response to return and
display an error if it comes back with a 500. Instead we need to make
sure that, no matter what, we always send a notification that the chart
build request has complete, successfully or otherwise.