Fix stack badge colors and tooltip spacing
Assign each stack a badge color depending on the status of the services
and their tasks. Fix spacing of tooltip for badge to remove unnecessary
line spacing.

Tweak CSS of Networks overview to make things a bit more presentable,
tidy, and clear. Add hover CSS to make it clear which aliases belong to
which service and network.